+++ src/Microsoft.AspNet.Security.Google/GoogleAppBuilderExtensions.cs
@@ -1,23 +1,17 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.AspNet.Security.Google;
-using Microsoft.Framework.DependencyInjection;
-using Microsoft.Framework.OptionsModel;
 using System;
+using Microsoft.AspNet.Security.Google;
+using Microsoft.Framework.OptionsModel;
 
 namespace Microsoft.AspNet.Builder
 {
     /// <summary>
     /// Extension methods for using <see cref="GoogleAuthenticationMiddleware"/>.
     /// </summary>
-    public static class GoogleAuthenticationExtensions
+    public static class GoogleAppBuilderExtensions
     {
-        public static IServiceCollection ConfigureGoogleAuthentication([NotNull] this IServiceCollection services, [NotNull] Action<GoogleAuthenticationOptions> configure)
-        {
-            return services.Configure(configure);
-        }
-
         /// <summary>
         /// Authenticate users using Google OAuth 2.0.
         /// </summary>
