+++ src/Microsoft.AspNet.Security.OAuth/OAuthAuthenticationExtensions.cs
@@ -2,9 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Globalization;
 using Microsoft.AspNet.Security.OAuth;
-using Microsoft.AspNet.Security.Infrastructure;
 using Microsoft.Framework.OptionsModel;
 
 namespace Microsoft.AspNet.Builder
