+++ src/Microsoft.AspNet.Security.Twitter/TwitterAuthenticationExtensions.cs
@@ -1,11 +1,9 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.AspNet.Security.Infrastructure;
-using Microsoft.AspNet.Security.Twitter;
-using Microsoft.Framework.DependencyInjection;
-using Microsoft.Framework.OptionsModel;
 using System;
+using Microsoft.AspNet.Security.Twitter;
+using Microsoft.Framework.OptionsModel;
 
 namespace Microsoft.AspNet.Builder
 {
@@ -14,11 +12,6 @@ namespace Microsoft.AspNet.Builder
     /// </summary>
     public static class TwitterAuthenticationExtensions
     {
-        public static IServiceCollection ConfigureTwitterAuthentication([NotNull] this IServiceCollection services, [NotNull] Action<TwitterAuthenticationOptions> configure)
-        {
-            return services.Configure(configure);
-        }
-
         public static IApplicationBuilder UseTwitterAuthentication([NotNull] this IApplicationBuilder app, Action<TwitterAuthenticationOptions> configureOptions = null, string optionsName = "")
         {
             return app.UseMiddleware<TwitterAuthenticationMiddleware>(
